+++ Samples/BasicApp/ClixNotificationExtension/NotificationService.swift
@@ -6,7 +6,7 @@ import UserNotifications
 class NotificationService: ClixNotificationServiceExtension {
   override init() {
     super.init()
-    register(projectId: ClixConfiguration.projectId)
+    register(projectId: ClixConfiguration.shared.projectId)
 
     NSLog("[NotificationService] Notification service initialized")
   }