+++ Sources/Core/Clix.swift
@@ -73,8 +73,9 @@ public final class Clix {
   static let version = ClixVersion.current
   static let shared = Clix()
 
-  /// Provides access to notification-related functionality in the Clix SDK
-  public static let Notification = ClixNotification.shared
+  #if !APPLICATION_EXTENSION_API_ONLY
+    public static let Notification = ClixNotification.shared
+  #endif
 
   // MARK: - Storage Keys
   private static let configKey = "clix_config"
@@ -93,8 +94,8 @@ public final class Clix {
   private init() {}
 
   // MARK: - Private Methods
-  private func setConfig(_ config: ClixConfig) {
-    self.storageService = StorageService(projectId: config.projectId)
+  private func setConfig(_ config: ClixConfig) async {
+    self.storageService = await StorageService(projectId: config.projectId)
     guard let storageService = self.storageService else {
       ClixLogger.error("Failed to initialize StorageService")
       return
@@ -151,7 +152,7 @@ public final class Clix {
   public static func initialize(config: ClixConfig) async {
     do {
       ClixLogger.setLogLevel(config.logLevel)
-      shared.setConfig(config)
+      await shared.setConfig(config)
       let deviceId = try await shared.get(\.deviceService).getCurrentDeviceId()
       let token = try await shared.get(\.tokenService).getCurrentToken()
       let device = await DeviceService.createDevice(deviceId: deviceId, token: token)
@@ -162,6 +163,10 @@ public final class Clix {
       try await shared.get(\.deviceService).upsertDevice(device)
 
       ClixLogger.debug("Clix SDK initialized with environment: \(environment.toString())")
+
+      #if !APPLICATION_EXTENSION_API_ONLY
+        ClixNotification.shared.initialize()
+      #endif
 
       await shared.initCoordinator.completeInitialization()
     } catch {
@@ -185,7 +190,7 @@ public final class Clix {
   }
 
   static func initialize(projectId: String) async throws {
-    let storageService = StorageService(projectId: projectId)
+    let storageService = await StorageService(projectId: projectId)
     let config = await storageService.getWithRetry(Self.configKey, fallbackValue: ClixConfig(projectId: projectId))
     await initialize(config: config)
   }